+++ joe/regex.h
@@ -6,5 +6,5 @@
  *	This file is part of JOE (Joe's Own Editor)
  */
 
-int escape(int utf8,char **a, int *b);
-int pmatch(char **pieces, char *regex, int len, P *p, int n, int icase);
+int escape(int utf8,char **a, ptrdiff_t *b);
+int pmatch(char **pieces, char *regex, ptrdiff_t len, P *p, ptrdiff_t n, int icase);
